bsps/powerpc: Address -Wsign-compare warnings
This warning occurs when comparing a signed variable to an unsigned one.
This addresses warnings that only occurred on 64-bit targets.  For the
ones which only appeared on 64-bit targets, the cause was frequently
a mismatch when comparing a combination off_t, ssize_t, and int.

+++ bsps/powerpc/beatnik/flash/flashcfg.c
@@ -162,7 +162,7 @@ struct bankdesc *b;
 			return 0;
 	}
 	/* bank is unsigned and cannot be < 0 */
-	if ( bank >= b->size ) {
+	if ( (uint32_t)bank >= b->size ) {
 		if ( !quiet )
 			fprintf(stderr,"Invalid flash bank #: %i; (too big)\n", bank);
 		return 0;
